+++ app/javascript/packs/application.js
@@ -7,11 +7,13 @@ import Rails from "@rails/ujs"
 import Turbolinks from "turbolinks"
 import * as ActiveStorage from "@rails/activestorage"
 import "channels"
-
 import "jquery";
 import "popper.js";
 import "bootstrap";
 import "../stylesheets/application"
+window.$ = window.jQuery = require('jquery');
+require('packs/raty')
+
 
 
 Rails.start()
