+++ app/views/admin/posts/index.html.erb
@@ -9,17 +9,30 @@
             <th>タイトル</th>
             <th>店舗</th>
             <th>エリア</th>
+            <th>評価</th>
           </tr>
         </thead>
         <tbody>
           <% @posts.each do |post| %>
           <tr>
-            <th><%= post.user.user_name %></th>
+            <td><%= post.user.user_name %></td>
             <th><%= link_to admin_post_path(post.id) do %>
                 <%= post.menu %></th>
                 <% end %>
             <th><%= post.shop.shop_name %></th>
             <th><%= post.area.prefecture %></th>
+            <th><div id="star-rate-<%= post.id %>"></div>
+                <script>
+                $('#star-rate-<%= post.id %>').raty({
+                  size: 20,
+                  starOff:  '<%= asset_path('star-off.png') %>',
+                  starOn :  '<%= asset_path('star-on.png') %>',
+                  starHalf: '<%= asset_path('star-half.png') %>',
+                  half: true,
+                  readOnly: true,
+                  score: <%= post.rate %>,
+                });
+                </script></th>
           </tr>
         </tbody>
         <% end %>
