+++ app/views/user/posts/new.html.erb
@@ -3,19 +3,19 @@
     <div class="col-12 mx-auto">
       <h4 class="post_new">レビュー投稿</h4>
       <%= form_with model:@post do |f| %>
-      <div class="form_group col-8">
+      <div class="form_group col-5">
         <%= f.label :menu, "タイトル" %>
         <%= f.text_field :menu, class: "form-control" %>
       </div>
 
-      <div class="form_group col-8">
+      <div class="form_group">
         <%= f.label :post_image, "投稿画像" %></br>
         <%= f.file_field :post_image, accept: 'image/*' %>
       </div>
 
-      <div class="form_group col-8">
+      <div class="form_group col-7">
         <%= f.label :body, "本文" %>
-        <%= f.text_field :body, size: "40x10", class: "form-control" %>
+        <%= f.text_area :body, class: "form-control", size: "60x10" %>
       </div>
 
       <div class="form_group">
@@ -26,6 +26,24 @@
       <div class="form_group">
         <%= f.label :area_id, "エリア選択" %></br>
         <%= f.collection_select :area_id, @areas, :id, :prefecture %>
+      </div>
+
+      <div class="form-group row" id="star"><!--id="star"でセレクタを指定している-->
+        <%= f.label :rate,'評価:', class:'col-md-3 col-form-label' %>
+        <%= f.hidden_field :rate, id: :review_star %>
+      </div>
+      <!-- 評価javascript -->
+      <div class="rate-group mx-auto">
+        <script>
+        $('#star').raty({ //#starは11行目で指定したクラス名を入れ、.ratyはjqueryで使用できる星評価機能のプラグインを使っています。
+          size     : 30,
+          starOff:  '<%= asset_path('star-off.png') %>',//asset_pathでassets/imagesに入っているファイルを指定しています。
+          starOn : '<%= asset_path('star-on.png') %>',
+          starHalf: '<%= asset_path('star-half.png') %>',
+          scoreName: 'post[rate]',//rateカラムに値を保存しています。
+          half: true,//★の半分の入力を行なっています。(0.5間隔の値になる)
+        });
+        </script>
       </div>
 
       <div class="button_field">
