+++ config/webpack/environment.js
@@ -8,6 +8,7 @@ environment.plugins.prepend(
   new webpack.ProvidePlugin({
     $: 'jquery/src/jquery',
     jQuery: 'jquery/src/jquery',
+    jquery: 'jquery/src/jquery',
     Popper: 'popper.js'
   })
 )
