+++ db/migrate/20220510090004_create_shops.rb
@@ -3,7 +3,7 @@ class CreateShops < ActiveRecord::Migration[6.1]
     create_table :shops do |t|
       t.string :shop_name, null: false
       t.string :address, null: false
-      t.integer :bussiness_hour, null: false
+      t.string :bussiness_hour, null: false
       t.boolean :shop_status, null: false, default: false
       t.timestamps
     end
