+++ db/schema.rb
@@ -99,7 +99,7 @@ ActiveRecord::Schema.define(version: 2022_05_11_014114) do
   create_table "shops", force: :cascade do |t|
     t.string "shop_name", null: false
     t.string "address", null: false
-    t.integer "bussiness_hour", null: false
+    t.string "bussiness_hour", null: false
     t.boolean "shop_status", default: false, null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
